Added initial support for 4th parameter image intended to be the recovery image for an AUE chromeOS device. This image can be used as a source to overwrite files in the chromium image after applying the updated recovery image.
This is the first step to increasing compatibility with AUE chromebooks
via the script. The next step will be to support N following parameters
listing files or folder to overwrite.

For example, after using pyro to chromefy an Acer C720 (peppy), the trackpad and
sleep mode do not behave correctly. Replacing the files in /etc/gesture
and /usr/share/power_manager on the chromefyd image with those from the peppy image allows those features to work correctly.

This can be done manually but the intention is to adapt the script to do this automatically.